refactored router tests
