+++ app/routes/Index.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import Index from "./routes/_index";
+import Index from "./_index";
 import "@testing-library/jest-dom"; // So we can use toBeInTheDocument assertion
 
 it("should show welcome message", () => {
